decouple marpa_m_test
